requirements.txt
py-cord==2.0.0b5
python-dotenv==0.20.0
python_dateutil==2.8.2
